+++ env.template
@@ -23,7 +23,6 @@ SCHEDULE_JOB=1 0 1 * * *
 SUBMISSION_DOC_URL=https://datacommons.cancer.gov/ 
 
 #Token settings
-TOKEN_SECRET=xxxxxx
 #set API token timeout to 1 day
 TOKEN_TIMROUT=86400
