+++ services/aws-request.js
@@ -1,0 +1,112 @@
+const AWS = require('aws-sdk');
+require('aws-sdk/lib/maintenance_mode_message').suppress = true;
+const {verifyApiToken,verifySubmitter} = require("../verifier/user-info-verifier");
+const path = require("path");
+const config = require('../config');
+
+const S3_GET = 'getObject';
+const S3_KEY = 'Key';
+const S3_SIZE= 'Size'
+const S3_CONTENTS = 'Contents'
+/**
+ * This class provides services for aWS requests
+ */
+class AWSService {
+    constructor(submissionCollection, organizationService, userService) {
+        this.organizationService = organizationService;
+        this.userService = userService;
+        this.submissions = submissionCollection;
+        this.s3 = new AWS.S3();
+    }
+    /**
+     * createTempCredentials
+     * @param {*} context 
+     * @param {*} submissionID 
+     * @returnsv {
+            accessKeyId: String
+            secretAccessKey: String
+            sessionToken: String
+        }
+     */
+    async createTempCredentials(params, context) {
+        //1. verify token and decode token to get user info
+        const userInfo = verifyApiToken(context, config.session_secret);
+        //verify submitter
+        await verifySubmitter(userInfo, params?.submissionID, this.submissions, this.userService);
+        //2. create temp credential
+        // Initialize an STS object
+        const sts = new AWS.STS();
+        const timestamp = (new Date()).getTime();
+        const s3Params = {
+            RoleArn: config.role_arn,
+            RoleSessionName: `Temp_Session_${timestamp}`
+        };
+        return new Promise((resolve, reject) => {
+            
+            sts.assumeRole(s3Params, (err, data) => {
+                if (err) reject(err);
+                else {
+                    resolve({
+                        accessKeyId: data.Credentials.AccessKeyId,
+                        secretAccessKey: data.Credentials.SecretAccessKey,
+                        sessionToken: data.Credentials.SessionToken,
+                    });
+                }
+            });
+        });
+    }
+    /**
+     * 
+     * @param {*} bucket s3 bucket name
+     * @param {*} prefix s3 bucket file prefix
+     * @returns file
+     */
+    async  getLastFileFromS3(bucket, prefix, filter){
+
+        const data = await this.s3.listObjects(getS3Params(bucket, prefix)).promise();
+        const files = data[S3_CONTENTS].filter(k=>k[S3_KEY].indexOf(filter)> 0)
+        if(files.length > 0)
+        {
+            const lastFile = files[files.length-1];
+            let key = lastFile[S3_KEY];
+            let fileName = path.basename(key);
+            let downloadUrl = await createDownloadURL(s3, params.Bucket, key);
+            let size = lastFile[S3_SIZE];
+            return {fileName: fileName, uploadType: uploadType, downloadUrl: downloadUrl, fileSize: size};
+        }
+        else return null;
+    }
+    
+    async  createDownloadURL(s3, bucketName, key) {
+        let expiration = (config.presign_expration && /^[0-9]*$/.test(config.presign_expration))? 
+            parseInt(config.presign_expration):3600; //defult value is 1 hour
+        const params = {
+            Bucket: bucketName,
+            Key: `${key}`,
+            Expires: expiration, 
+        };
+        return new Promise((resolve, reject) => {
+            s3.getSignedUrl(S3_GET, params, (error, url) => {
+                if (error) {
+                    reject(error);
+                } else {
+                    resolve(url);
+                }
+            });
+        });  
+    }
+    
+}
+
+function getS3Params(bucket, prefix){
+    return {
+        Bucket: bucket,
+        Delimiter: '/',
+        Prefix: `${prefix}/`
+    };
+}
+module.exports = {
+    AWSService
+};
+
+
