+++ utility/string-util.js
@@ -53,6 +53,23 @@ const extractAndJoinFields = (data, fieldsToExtract, splitter = ",") => {
     );
 }
 
+async function getSubmisssionRootPath (submission, organizationService) {
+    let rootPath = submission.rootPath;
+    if(!rootPath){
+        let orgId = null;
+        if(typeof submission.organization != "string"){
+            orgId =submission.organization._id
+        }   
+        else{
+            const org = await organizationService.getOrganizationByName(submission.organization);
+            orgId = org._id;
+        }
+            
+        rootPath = `${orgId}/${params.submissionID}`;
+    }
+    return rootPath;
+}
+
 module.exports = {
     isCaseInsensitiveEqual,
     isElementInArray,
@@ -61,5 +78,6 @@ module.exports = {
     getUniqueArr,
     parseArrToStr,
     replaceMessageVariables,
-    extractAndJoinFields
+    extractAndJoinFields,
+    getSubmisssionRootPath
 }
